QandAFROMCSV/requirements.txt
langchain
langchain-google-genai
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
sentence-transformers
InstructorEmbedding